This is a fix for NCL-681.
uv2vr_cfd and uv2dv_cfd were only returning a missing value if "u" had
a missing value set. It is possible for the output to contain missing
values even if none of the input values are missing, so I changed this
behavior to a missing value is always returned.

If no missing values are input, then the default missing value for the
type of "u" is used. v's missing value doesn't come into play at all.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_5_2_1@11589 10fe2890-8a40-4696-8b1d-7ccbb19c398e