This fixes NCL-686 and PyNIO-31.
Normally for statistical variables in GRIB2 files, NIO adds the statistical process duration to the  
forecast time and gives the adjusted forecast time for the end of the statistical period. This is 
based on a footnote of the relevant template documentation that says the forecast time plus the initial
time defines the beginning of the statistical period. However, because this would result in some forecast
times becoming negative in some instances for NDFD data, and the forecast time is defined as an unsigned
number (by WMO I think), NDFD chooses to make the forecast time the end of the statistical period. The
NIO library has been adapted to recognize this, hopefully without screwing up any other GRIB2 type files.
I have tested with all the GRIB2 files I have access to and so far so good.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_5_2_1@11593 10fe2890-8a40-4696-8b1d-7ccbb19c398e