Mary's latest encounter with the ARSCAM/ARPRAM bug revealed a special case to be accounted for. Specifically, when tracing along hole-connectors in phase 6 of the algorithm, the top node of the connector splits an original segment, and because the geometry may get shifted slightly in the process, we want to use the coorinates of the original segment in performing angle tests, not the coordinates of the inserted node. All good and well -- we detect these instances by the fact that their node number is greater than ILW, but their next or prev pointers point to a user segment (i.e., a node number less than ILW).
It turns out that the first node of the first hole-connector that gets inserted is at the 
bottom of the segment (recall that bottom nodes never cause us grief, only top nodes), and
its previous pointer points back to a user segment. Thus is looks like we are dealing with
a top node. This first node has an ID of exactly ILW+1 (its the first inserted temp. node),
so we can exclude on this basis.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_5_2_1@11628 10fe2890-8a40-4696-8b1d-7ccbb19c398e