Cleaned up warnings from -Wall.
Fred and I talked about the warnings in these two files, and 
ended up changing a couple of "int" functions to "void", and 
cleaning up some fprintf statements that had an unneeded "itmp"
argument.

We still get warnings for both ps.c and pdf.c:

     warning: format not a string literal and no format arguments

but we're going to leave these for now.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11777 10fe2890-8a40-4696-8b1d-7ccbb19c398e