NCL-755: inclusion of Aaron Maus's gridspec code on an optional basis
This implements conditional inclusion of interfaces to the gridspec nccf level routines.
It requires NetCDF 4.1.1 or later built with the '--with-libcf' option. An option has been added
to the Configure script to set the BuildGRIDSPEC define. It defaults to 0, don't build.

I modified the code to make it ng_size_t compatible, eliminate -Wall messages, and correct a few 
minor issues.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11779 10fe2890-8a40-4696-8b1d-7ccbb19c398e