NCL-755: optional recognition of interface to gridspec nccf routines.
Added a new question concerning experimental support for an interface to 
the gridspec nccf routines, as developed by Aaron Maus. The prerequisite
is NetCDF 4.1.1 or later built with the '--with-libcdf' option. The CF
gridspec code is definitely alpha quality. There are several header files
from the gridspec libsrc directory that need to be accessible to allow the
nfp nccf.. routines to build. They do not get installed automatically and
so must be copied to the installation include dir. These are mpp.h,
mpp_io.h, and mpp_domain.h. I have been testing the code and the errors I 
have encountered so far are in the gridspec libsrc, not in Aaron's code.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11781 10fe2890-8a40-4696-8b1d-7ccbb19c398e