Mary sent the following:
===  16 Sept 2010
Dennis,

There's a minor bug in PopLatLon in the copying of metadata from x to Xout.

xmsg is ending up double (because one of the input arrays is double), but xOut@missing_value is float, and hence when it tries to write xmsg into xOut@missing_value, you get a fatal error.

===
I found only one instance where I explicitly created xOut@missing_value.
Just delete an reassign ... this way the type of '@missing_value' will be the type of 'xmsg'

  if (isatt(xOut,"missing_value")) then
      delete(xOut@missing_value)         
  end if
  xOut@missing_value = xmsg




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11794 10fe2890-8a40-4696-8b1d-7ccbb19c398e