This addresses NCL-805. Checks for INT_MAX should be done earlier in code.
Some of the INT_MAX checks were inside do loops, which is expensive.

Some of the error statements had the wrong function names, and weren't
necessarily printing the correct dimension value.

Also changed dz_height wrapper so that double is returned if *any*
of the input is double (before, it just checked zsfc, which was 
inconsistent with documentation).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11886 10fe2890-8a40-4696-8b1d-7ccbb19c398e