NCL-854: this fixes a fairly large memory leak that showed up in the plotting script countrydiv.ncl.
This script uses the map dataset "Earth..4". It turns out that it was not fully recognized by the
data handler, and therefore took a path that caused some memory not to be freed. Now this should
no longer be a problem for any dataset including user-defined sets such as some of the paleo 
datasets. I also had to modify the code to recognize that Earth..4 has different names for some
inland water bodies. Lakes formerly named like this: "Lake Texoma (North America)" are called
simply "Lake Texoma" in "Earth..4". They are part of the group "North American Lakes".  
Practically speaking this means that scripts like mediumres_map.ncl need to be modified to
work with "Earth..4": the entity "lake texoma (north america)" needs to change to "lake texoma".



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11930 10fe2890-8a40-4696-8b1d-7ccbb19c398e