These do exactly the same thing ... calculate a pattern correlation.
The only difference is the the patternCor3.f does the looping
over multiple grids in the fortran code rather thain in the C-wrapper.
Perhaps it is faster. Mary can coose which one.

TEST doc is at:

http://test.www.ncl.ucar.edu/Document/Functions/Built-in/espatccr.shtml

I am not sure yet wht I will call the function

espatccr or esccr_pat or ???






git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@11950 10fe2890-8a40-4696-8b1d-7ccbb19c398e