Addresses NCL-620: Improve naming of PNG filenames.
Performed some refactoring/consolidation of how filenames are
generated for all formats. For PNGs (and any other future format that
does not support multiple images/pages per file), the general scheme
has been to append a sequence number to the filename root. However,
now if only one file/plot is generated, this is skipped and the
resulting filename is just the root as provided by the user, along
with an appropriate suffix (e.g., ".png").



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11951 10fe2890-8a40-4696-8b1d-7ccbb19c398e