NCL-854: Fixed an uninitialized variable reference in the nfp functions. This was uncovered using valgrind on the wrf_surface3.ncl script in the plotting test directory.
The wrf_smooth_2d routine was passing the missing_a value retrieved from
the 'fld' parameter to the Fortran filter2d procedure without checking to 
see if 'has_missing_a' was non-zero. Now missing_a is replaced with the default
float or double missing value if has_missing_a is 0.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11953 10fe2890-8a40-4696-8b1d-7ccbb19c398e