NCL-885: isfilevar was not working with a file list element.
The problem was that in implementing a more universal implementation of the 'list' type, Wei added some code that
changed the 'obj_type_mask' field of objects added to a list via ListPush. Neither Wei nor I know why this was
a desirable thing to do, but it may become evident when the use of the 'list' type becomes more widespread. The
result was that functions that take a file as a parameter did not see the file list element as a valid input type.
I have commented out the lines that Wei added, fixing the immediate problem. The lines are still there in case
they might prove useful later on.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11958 10fe2890-8a40-4696-8b1d-7ccbb19c398e