This check in related to NCL-842:
Because the way we define Ncl_Typeint, etc in NclDataDefs.h,
the current NCL_NUMERIC_TYPE_MASK and NCL_ENUMERIC_TYPE_MASK
are not exclusive, which results:

data_type & NCL_NUMERIC_TYPE_MASK
and
data_type & NCL_ENUMERIC_TYPE_MASK
are not unique, which cause problem when we check if a data
is NUMERICAL or ENUMERICAL.

The patch will fix the above problem, by additional checks.


The reason to cause this problem is that we have too-many "types"
in NclDataDefs.h to define, where we use enumeration to define
each type. Dave and I have checkeed, as in ANS C,
enumeration is type "int", which means we can only "define"
up to 32 types, "bit" unique. (But in C++, one can use any integer type.)

It seems we can use "long long" to define these variables,
as on Mac, it seems to work, but we are not sure if it works on
other systems.

We have checked our NCL code, and it seems there is not so many
places use these two MASKs. We decided to fix this problem the
current way:
1. Do the mask operator first.
2. Pick those overlapped types.
3. Treat those overlapped types seperately.

Sugestions will certainly be exampined.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11960 10fe2890-8a40-4696-8b1d-7ccbb19c398e