NCL-55: this fixes the HDF4 code to handle unsigned integers both for reading and writing. Also along the way: changes to getfilevartypes to recognize the enumeric types. Minor changes to error output in NclFile.c and a fix for an uninitialized memory reference in NclHDFEOS5.c.
Changes to ncl_convert2nc will be forthcoming as it needs modification to handle the enumeric
types -- which need to be converted to NetCDF classic model types for writing to NetCDf.
Also note that ncl_filedump (actually write_matrix) barfs when asked to output unsigned or 8-byte
integers data variables (using the -v option), and also strings.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11978 10fe2890-8a40-4696-8b1d-7ccbb19c398e