NCL-897: Fixes ncl_convert2nc to convert unsupported types to those supported in classic model NetCDF
ushort converted to short
uint  converted to integer
ulong converted to integer if 32 bits.
64bit integers are converted to double
strings are converted to character arrays -- this requires the addition of a new dimension, ncl_strlen_0, ncl_strlen_1, etc.
The dimension size is of the longest string. The trailing null byte is *not* written to the NetCDF.
A dimension is created for each character array even if happens to have the same length as an existing array.
An added attribute called 'ncl_converted_from_type' specifies the original type.
A warning is issued for each conversion, e.g: 

Classic model NetCDF does not support string types, converting CATEGORY to a character array
Dimension 'ncl_strlen_0' will be added

This behavior takes place by default but there are some options to control it:

  [-no-uc]              turn off default conversion of unsigned integer types to corresponding signed types
                        -- unsigned variables in the input will be skipped
  [-no-i64c]            turn off default conversion of 64-bit integer types to double
                        -- 64-bit integer types in the input will be skipped
  [-no-sc]              turn off default conversion of string types to character
                        -- string types in the input will be skipped
  [-no-aa]              turn off default addition of the 'ncl_converted_from_type' attribute
                        when variable types are converted

I am not certain that converting the large integer types to double is the best policy, but it is generally a moot point for now,
since we don't have any examples to try. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11980 10fe2890-8a40-4696-8b1d-7ccbb19c398e