NCL-14,NCL-905: Changes default fill values.
This code changes the default fill values to the same values used by the NetCDF library (except for string types).
Also it adds a command line option '-o' (for 'old-behavior') that returns the default fill values to their NCL 5.x values.
Additionally it adds a new function 'default_fillvalue(string : type)' that returns the current default fill value for
a particular type; plus a new procedure 'set_default_fillvalue(string : type, value)' that sets a new default fillvalue.
The value must be promotable to the specified type. This function can also be used to set all the fillvalues to a
particular set of defaults by substituting "all" for the type argument, and one of "ncl_v5", "ncl_v6", or "default" for
the value argument. "ncl_v6" and "default" are synonyms. Note that if the -o option is used, the default fill values
are set prior to loading any default libraries or scripts and prior to executing any command line assignments.

I had to change the scanf lines in ncl.l to use "%llu" rather than "%lld" for the input conversion. Hopefully this will
work on all systems, but it seems a little clunky. If it doesn't work everywhere, we will have to have separate flex code to handle
the unsigned long long type. Using %lld it is not possible to get a number bigger than the maximum signed long long value.

Also I fixed the problem with incorrect printing of byte types (now they print as decimal integers -- not hexadecimal, which seem
to work only for unsigned values).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11987 10fe2890-8a40-4696-8b1d-7ccbb19c398e