NCL-920: Changes to make toxxxx functions produce results same as from yyyytoxxx. Added: Register _NclIinttochar as "inttocharacter" Register _NclIchartoint as "charactertoint"
Added: _NclIshorttoint as "shorttoint"
and then:
Register _NclIshorttoint as "shorttointeger"

There still  problem with floattoint and toint(with float as input),
when float number is: 2147483647,
which equals INT_MAX, which floattoint, and toint
both have problems to convert to correct int on my mac
(it actually converted to -2147483648, which is INT_MIN).

There are few commented lines left there need to be cleaned.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@12002 10fe2890-8a40-4696-8b1d-7ccbb19c398e