This addresses a problem that WRF users have with zooming in plots.
The wrf_overlays and wrf_map_overlays functions always set
tfDoNDCOverlay to  True, which means if you want to zoom
in on the plot, then you have to know how to subset the
data correctly.

I added an internal resource, "LatLonOverlay", that if set to True,
will cause tfDoNDCOverlay to *not* be set. This is assuming that
the user has already set sfXArray and sfYArray to the XLONG/XLAT
arrays.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@12022 10fe2890-8a40-4696-8b1d-7ccbb19c398e