NCL-32: re-opened in response to ARSCAM/ARPRAM failure found by Adam Phillips.
Code had previously been added to Phase 6 of both ARSCAM & ARPRAM to
deal with situations where line crossings get introduced due to
numerical imprecision; when lines are split to add hole-connectors,
the new node may not lie exactly on the original line, and thus the
end points of the split line may get shifted up or down
slightly. (This is the code related to tracking the variable
"IFRMXY").

That same code was needed for phase 5 as well, of both ARSCAM & ARPRAM.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@12052 10fe2890-8a40-4696-8b1d-7ccbb19c398e