NCL-981:
Some work for list convert->local and changes the
behavior of ListGetType.

Where if we create list with: lst = [/x, y, z/],
later, when we want to check index of something,
convert->local has some trouble to compare dimensions.
For list, we choose to skip this step for now.

For ListGetType, now it returns the type(s) have set.
Where previous, it always returns two strings.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@12072 10fe2890-8a40-4696-8b1d-7ccbb19c398e