NCL-992: Fix for problem with the GRIB2 parameter_discipline_and_category attribute.
What I actually did was improve the Grib2ReadCodeTable routine to be able to select on both
the numbers at the beginning of each line of the code table files. In almost every case
the 2 numbers at the beginning (separated by the ':' character) are the same, but 4.2.table
contains different values. This changes looks at both numbers allowing the correct line to
be selected based on matching both of them when relevant.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12096 10fe2890-8a40-4696-8b1d-7ccbb19c398e