Added a test gc_inout_mask routine.
This routine was an attempt to make examples like mask_9.ncl run
faster, by doing the masking inside a C code instead of an NCL code.

Unfortunately, it doesn't appear to be any faster than the straigh
gc_inout method. I think this is because of all the data copying
necessary for coercion.  More testing is required.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12102 10fe2890-8a40-4696-8b1d-7ccbb19c398e