NCL-989: fix for problems with do loops with mixed types.
This fix regularizes the behavior for do loops with mixed types for the loop counter variable,
the start expression, the end expression, and the stride value.
It also checks for overflow and stops the loop when overflow is encountered.
It modifies the loop test algorithm to ensure that overflow only occurs when the max or min
value of the loop counter type is actually exceeded.

Mixed type usage is based on standard NCL coercion rules as follows:
The start expression must be coercible to the type of the loop increment variable when
the increment variable is defined prior to the beginning of the loop. The stride value, when
specified, must also be coercible to the type of the loop increment variable. There is
no coercibility requirement between the loop increment variable and the end value. But it
should be noted that the test for loop completion relies on converting both the increment variable
and the end value to double. This means currently that tests for completion involving int64 
or 8 byte longs with more than approximately 14 decimal digits may not work properly due to insufficient
precision in the double type. This may need to be revisited but I have left out an alternate 
scenario in an effort to keep the looping code as simple as possible.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12104 10fe2890-8a40-4696-8b1d-7ccbb19c398e