NCL-1006: now recognizes the ubyte type
I modified ncl_convert2nc to eliminate the int8 and uint8 types and recognize the ubyte type.
However, rather than converting ubyte to character in NetCDF it now converts it to byte. I am
thinking that this is more in keeping with the conversion for the other unsigned integer types,
and when reading back the NetCDF, at least from NCL, it is simpler to convert from signed to 
unsigned when necessary instead of from character to ubyte. But feel free to argue this point.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12117 10fe2890-8a40-4696-8b1d-7ccbb19c398e