NCL-1007: a supporting fix to help ncl_filedump handle new numeric types
Since write_matrix is written in Fortran it does not handle unsigned types 
(at least that I know of). I have therefore modified the 'sprinti' function
to handle all integer types. This is used in ncl_filedump to help generate
properly spaced output for printing types that write_matrix cannot handle.

Also fixes a problem with getfilevartypes: it was not handling int64 or
uint64 types.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12127 10fe2890-8a40-4696-8b1d-7ccbb19c398e