NCL-1007: ncl_filedump now can print out the values of any supported NCL numeric type.
I wrote a little NCL procedure that prints out the values of types that write_matrix
cannot handle. It is considerably slower than write_matrix however, so for ubyte and
ushort I up convert to short and integer respectively and hand them off to write_matrix.

The NCL procedure relies on the changes I made to sprinti, now supporting any integer type.

I also modified the calculation of how many items to print on a line, simplifying it 
quite a bit. However, in a few cases it may mean that the number of columns is not
the same as before. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12128 10fe2890-8a40-4696-8b1d-7ccbb19c398e