NCL-1005: fix (hopefully) for segfault with missing files read with addfiles.
The variable aggregation code was missing a check for file existence in the "join" mode
brank of the code. I do not have access to the only machine where this problem
manifested so I have not been able to test the fix other than to ensure that 
it works with the test code on my machine (and that all the regular addfiles tests
still run).




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12135 10fe2890-8a40-4696-8b1d-7ccbb19c398e