NCL-1035: fix for problem with Andy Mai's script:
There were 2 independent problems:
1) if the type argument to the new operator was input as a string, the
temporary object created was not being released properly after use. This
accounted for the most growth in the object table for the Mai script.
2) the dimension name assignment operation was not properly freeing temporary
strings assigned using variable subsections, such as:
x!0 = var(0)
In this instance a temporary VARSUBSEL type variable is created and passed to
the dimension name assignment operator. But the code was only releasing objects
when they represented literal values. In normal variable to variable assigment
there is an operator that takes both the from and to variables as a whole along
with their subselection parameters. No subselection variable is created because
everything is handled within the routine. 

There may be other instances of this type of problem in the Execute.c code. I think
it deserves fuller study, but because it requires careful testing to make sure that
only things that should be freed are freed, I am postponing further work on this 
until the other 6.0.0 tickets are taken care of. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12164 10fe2890-8a40-4696-8b1d-7ccbb19c398e