NCL-1046: fix for undefined order of loading for scripts in NCL_DEF_SCRIPTS_DIR
Scripts are now loaded in ascii case-sensitive order (numbers before letters, upper case before lower case).
I also changed the behavior so that files that do not end with '.ncl' are simply ignored. Previously
a fatal error was generated for each file without a proper suffix. A warning message is generated if
there are no valid scripts in the directory, and a script that fails to load still generates a fatal
error.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12180 10fe2890-8a40-4696-8b1d-7ccbb19c398e