NCL-648: this fixes the problem with CMC GRIB2 data
The same problem as the GRIB1 files had is now fixed for GRIB2. The problem
was global grids specified with the same longitude value for Lo1 and Lo2.
This is actually according to the GRIB2 specs where all longitudes must be
specified with positive values. But we can't do this without generating 
non-monotonic longitudinal coordinate arrays. So in this case 360 is 
subtracted from Lo1 to achieve the right results. Fixed a couple of 
warning messages in the other 2 files.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12199 10fe2890-8a40-4696-8b1d-7ccbb19c398e