A change was made between 5.2.0 to 5.2.1 to accomodate 'long' integers. Marco suggested the following change.
I incorporated Marco's change. It worked on my simple test.

MH ... can u look when u get a chance?

===
Hi

The update from NCL 5.2.0 to 5.2.1 applies the following change to the
function readAsciiTable() [ncarg/nclscripts/csm/contributed.ncl line 7893]:

-      head = readAsciiHead( fName, opt(0))
-      nh   = dimsizes(head)                 ; # header records
+      if(opt(0).gt.0) then
+        head = readAsciiHead( fName, opt(0))
+        nh   = dimsizes(head)                 ; # header records
+      else
+        nh = 0                                ; no header to read
+      end if

According to the documentation opt(0) can be a number or a string to
specify the end of the header. In 5.2.1 the function does not work
correctly anymore if opt(0) is a string.

To fix this I changed the if-statement in the code above to:

   if((nopt.eq.1 .and. typeof(opt).eq."string") .or. opt(0).gt.0) then

With that change my scripts work again.

Cheers,
Marco
-- *************************************** Marco Steinacher Climate and Environmental Physics Physics Institute, University of Bern Sidlerstrasse 5, CH-3012 Bern 


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12216 10fe2890-8a40-4696-8b1d-7ccbb19c398e