NCL-1073: VectorPlot not handling high res data overlaid on small map areas.
This is a partial fix for this problem. It fixes vectors drawn by the LLU vvectr code.
Basically it modifies the vector position and angle code to use double precision and
avoid conversion to NDC coordinates until the coordinates are fully determined in user
space. This would not have been possible when VectorPlot was originally developed because
the double precision map routines were not avaiable then.

Curly vectors and streamlines, both of which are rendered by the LLU stream routine
are not yet fixed for this problem. It is a bit more complicated because the code is
split up into several functions and it will probably require changes to the LLU code to fix.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12231 10fe2890-8a40-4696-8b1d-7ccbb19c398e