NCL-1053: recognize when coordinate arrays are actually linear
This fix allows tr[XY]Log to be set for plots where coordinate arrays
are specified but the coordinates are actually linear. Formerly, this
would result in an error because the presence of coordinate arrays caused
the code to assume the coordinates are irregular. This could cause some
subtle differences in plots where now the LogLinTransObj is used instead
of the IrregularTransObj, but I hope it will not be generally noticeable.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12252 10fe2890-8a40-4696-8b1d-7ccbb19c398e