NCL-1109: fix for abort on or after 'stop record'
There was a flag that gets set 'rec' indicating that recording is on; but it 
was not being turned off when the record is stopped. As near as I can tell
this error has always been in the code -- at least as far back as r5864 it was there.
The process was continuing to try to write to a file that had been closed.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12292 10fe2890-8a40-4696-8b1d-7ccbb19c398e