NCL-869: Fix for slow asciiread
[Fixed by Dave, merged to 6.0.0 branch by Mary.]

The problem was long lines of text each of which gets inserted into
the Quark table. The Quark code was coded with the assumption that
strings would be short (i.e. 32 characters or less). The "signature"
generator in NrmStringToQuark is supposed to generate a more-or-less
unique integer for whatever string is applied. The problem was the 
left-shifting algorithm was removing all but the last 32 characters
from affecting the integer signature. This is resolved by modifying
the algorithm to borrow a bit from the SHA1 algorithm: now the sig
generated from each 32 characters are all added together, resulting
in a number based on all characters in the string. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/tags/ncl_6_0_0@12323 10fe2890-8a40-4696-8b1d-7ccbb19c398e