NCL-1133: Fix for incorrect fill of areas surrounding data in certain cases.
[This is Dave's fix.]

The seam.ncl problem was easy: I needed to change the default transparent out-of-range value
to the background color for raster plotting.

In the other case, it has to do with the fact that the triangular mesh contouring routine 
does not distinguish between outside the grid boundaries and missing value areas. My fix
which may be slightly controversial is to make the default cnMissingValFillColor transparent
rather than background color and also change the default cnMissingValFillPattern from 'hollowfill'
to 'solidfill'. Since these are switched the default is still for no fill in missingval areas.
It should also have very little effect on existing codes, and actually makes it easier to turn
on the missing val fill, because you no longer have to set the pattern if you just want solid
fill -- setting the color will be enough. I looked through all the example scripts that use
these resources and I believe that there will be no visible effect from changing these defaults.
Hope I'm right.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/tags/ncl_6_0_0@12325 10fe2890-8a40-4696-8b1d-7ccbb19c398e