NCL-1149: fixes problem w/ blank area in plot on one size of grid boundary
This problem was introduced with the implementation of the cnGridBoundFillColor
resource. But it stems from the Areas code for resolving contradictory area id information.
The algorithm is to simply use the area id that occurs most often within an area, but if 
the ids occur equally often the lowest number was being used. I changed this to use the 
highest id number on the theory that the "regular" fill areas all have higher numbers than
the "special" areas (at least by default) and therefore the "regular" areas should be
preferred. This single letter change ("GT" -> "GE") fixed this problem and does not 
seem to result in any visible difference for any of the test plots.  



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12419 10fe2890-8a40-4696-8b1d-7ccbb19c398e