NCL-1155: Fix for problems reading files with single dimensions that exceed 2^31.
This was partly a problem with using the old NetCDF2 ncdiminq function and partly a problem with the NetCDF reading NCL code. 
I replaced ncdiminq with nc_inq_dim, the NetCDF 3 and later function. 
Similar code was in the HDF4 reader and replaced. The change to the yMakefile was actually to help incorporate Wei's
recent changes into PyNIO.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12478 10fe2890-8a40-4696-8b1d-7ccbb19c398e