NCL-1159: fix for problem of logical assignment where the left side operator is the same as the left side of the assignment
This situation only arises because of the special code in Execute.c that implements 'short-circuit' evaluation of scalar logical
values. Normal assignment does not occur in this case -- the left side operator value simply becomes the assignment value. I
added a bit of code in WriteVar that recognizes this condition and eliminates the error.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12492 10fe2890-8a40-4696-8b1d-7ccbb19c398e