Added a new procedure, gsn_draw_coordinates.
  gsn_draw_coordinates(wks,plot,data,res)

I've wanted this for a looong time. It simply draws the locations of
the X/Y coordinate values (as markers) associated with the input
"data" array on the given "plot".

If there are no coordinate arrays attached to "data", it will look for
the special lat2d, lon2d attributes.

If lat2d/lon2d or coordinate arrays don't exist, then you can attach
special xcoord/ycoord attributes to "res".

The following attributes are recognized, but the names might change:

  gsnDrawCoordsMissingColor - Wherever "data" is missing, the coords
                              will be drawn in this color. Default is
                              to draw all markers in foreground color.

  gsnDrawCoordsNonMissingColor - Wherever "data" is non-missing, the
                                 coords will be drawn in this color.

  gsnDrawCoordsAttach - By default, the markers are drawn with
                        gsn_polymarker. If this resource is True, then
                        they are attached with gsn_add_polymarker.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12617 10fe2890-8a40-4696-8b1d-7ccbb19c398e