NCL-1204: implements '-u' option without specified dimension name
This allows ncl_convert2nc pick the time-related dimension to use if no name is specified
along with the -u option. The -U option still allows a new name for the dimension, and it
works along with the -itime option if single-element dimensions are desired.
If more than one dimension would qualify, the dimension with the most elements is chosen. For GRIB the
dimension name must include "initial_time" as a substring. Otherwise, only the substring "time" (any case)
is required. Errors will result if a dimension that is not the left-most dimension is chosen by the algorithm.

Other fixes: a problem with printing character arrays in ncl_filedump was fixed.
The "conventions" attribute added by ncl_convert2nc was changed to "Conventions"



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12627 10fe2890-8a40-4696-8b1d-7ccbb19c398e