Adding Mohammad's original sparse matrix multiply.
This is so we can test his original scripts without any changes.

A version of this function was already added as "sparse_matrix_mult",
but the interface is different, and the row,col arguments may have
different meanings (i.e. 0-based instead of 1-based, possibly, I have
to look into this).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12653 10fe2890-8a40-4696-8b1d-7ccbb19c398e