Added another argument to sparse_matrix_mult that contains the output dimensions.
The only assumption you can make about  the size of the output array
is that the number of columns is the same as x's columns.

You can't assume anything about the number of rows.  So, I've
decided to add a fifth argument that contains the output dimension
sizes (not including the leftmost dimensions).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12672 10fe2890-8a40-4696-8b1d-7ccbb19c398e