NCL-1161: Fixes problem with using ncl_convert2nc
ncl_conver2nc /glade/data02/dsszone/ds627.0/ei.oper.an.ml/198901/ei.oper.an.ml.regn128uv.1989013118.grb
was not working because of an error in the csh scripting that shows up under these conditions:
1) the path is more than just the file (i.e. the file is not in the current working directory)
2) the actual file does not have the extension (e.g. '.grb' in this case)

I think this error has been around from the beginning of this script. 

The other problem noted in this JIRA ticket was fixed when I added the capability for the script to figure out the 
time dimension name if not specified when using -itime. Basically it was choking when trying to convert single element
string variables to netcdf char variables.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12674 10fe2890-8a40-4696-8b1d-7ccbb19c398e