NCL-1276: fix for areas left out near the poles in a raster plot using triangular mesh
The problem involved an interaction between checking the longitudinal and the latitudinal boundaries. 
The longitude was incorrectly being tested within the wrong cyclic period, but the test only occurred if
the latitude was just outside its valid range (but within range when subjected to a more accurate check that
accounts for floating point error).
I am slightly worried about this fix, because it simply involved uncommenting some code that was already in place, suggesting
that there may have been a reason to remove it. However, it seems right to me now, and I hope that the reason it was 
removed is no longer valid. I ran all the plotting tests and didn't notice any problems, but if any strangeness shows up in
triangular mesh plots, I will need to take another look. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12759 10fe2890-8a40-4696-8b1d-7ccbb19c398e