Cleaned up the "copy_VarCoords_n" and "copy_VarCoords_not_n" procedures.
This addresses NCL-742.

They both basically did the same thing: copy coordinate variabes from
one array to another, but skipping the given dimensions. The only
difference was that one expected both arrays to be the same size, and
the other wanted the second array to be a subset in size of the first
array.

I combined these two procedures into just one called "copy_VarCoords_not_n".

I then created one that does the opposite: it copies the coordinates
of the given dimensions from one variable to another.

Neither procedure requires that the given dimensions be monotonically
increasing. The only requirement is that they represent valid 
dimension in the "var_from" variable, and that the corresponding
dimension that you are copying to is the appropriate size.

The "var_to" variable will be copied to starting with the leftmost
dimension and going one dimension at a time.

git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12770 10fe2890-8a40-4696-8b1d-7ccbb19c398e