NCL-1276: backing out this change
As I feared uncommenting a piece of code in this script had the effect of causing what appears to be an endless loop in some of 
the ocean diagnostic files when cnLineLabelOn is true and the label type is "constant". Also the output ps file was growing without
bounds. I had made 2 changes to fix NCL-1276.  This was a change that I thought improved the situation but it was not necessary for the NCL-1276 fix. Still needs study.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12773 10fe2890-8a40-4696-8b1d-7ccbb19c398e