Moved some index remapping code into ESMF_regrid.
Originally, if you were dealing with unstructured data that had missing values
in the lat/lon arrays, you had to cull out these values do the interpolation,
and then remap back on to the new grid.

Now, the remapping back on to the new grid is done for you inside of
ESMF_regrid, if you set some attributes: RemapIndex, Indexes, and IndexesDims.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12779 10fe2890-8a40-4696-8b1d-7ccbb19c398e