NCL-1174: fixed problem with areafill colors.
I changed the default value of the resource cnFixFillBleed to True. This fixes the problem of
areafill colors not being fully contained inside the contour fill areas they belong with.
FYI I found the code that demonstrated a problem with cnFixFillBleed set to True
and verified that it no longer has a problem. Hopefully there will soon be no need for
the cnFixFillBleed resource.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@12785 10fe2890-8a40-4696-8b1d-7ccbb19c398e